Rename less files to make ressources compilation work
main.less is renamed to bundle.less and custom.less is renamed to
main.less. These changes were neccessary to make the ressources
compilation run without error. In the reference repo
collective/plonetheme.tango are the files named like they are now
in the text. So the training description gets closer again what is
done in the reference repo.