Update index.md change term Teaching to Giving a Training
Teaching implies a in my opinion outdated understanding of top down education. Some thing they cannot teach. Some hate teachers. The actual content of that pages reflect a more modern  aproach and are more inviting. This should be also present in the wording of the TOC and Headlines.

The content can also be improved to a more encouraging onboarding of trainers beyond it is already trying tis.